Separate out celoStateAtBlock into celo-specific file
